chore: move flaky dotMemory/dotTrace tests to ManualRunning project
