+++ tests/BenchmarkDotNet.IntegrationTests.ManualRunning/DotMemoryTests.cs
@@ -5,21 +5,21 @@ using System.Linq;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Configs;
 using BenchmarkDotNet.Detectors;
-using BenchmarkDotNet.Diagnostics.dotTrace;
+using BenchmarkDotNet.Diagnostics.dotMemory;
 using BenchmarkDotNet.Jobs;
 using BenchmarkDotNet.Portability;
 using BenchmarkDotNet.Toolchains.InProcess.Emit;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace BenchmarkDotNet.IntegrationTests
+namespace BenchmarkDotNet.IntegrationTests.ManualRunning
 {
-    public class DotTraceTests : BenchmarkTestExecutor
+    public class DotMemoryTests : BenchmarkTestExecutor
     {
-        public DotTraceTests(ITestOutputHelper output) : base(output) { }
+        public DotMemoryTests(ITestOutputHelper output) : base(output) { }
 
         [Fact]
-        public void DotTraceSmokeTest()
+        public void DotMemorySmokeTest()
         {
             if (!OsDetector.IsWindows() && RuntimeInformation.IsMono)
             {
@@ -29,9 +29,8 @@ namespace BenchmarkDotNet.IntegrationTests
 
             var config = new ManualConfig().AddJob(
             [
-                Job.Dry.WithId("ExternalProcess")
-                // TODO: Add InProcessEmitToolchain job when flaky test issue is resolved https://github.com/dotnet/BenchmarkDotNet/issues/2950
-                // Job.Dry.WithToolchain(InProcessEmitToolchain.Default).WithId("InProcess")
+                Job.Dry.WithId("ExternalProcess"),
+                Job.Dry.WithToolchain(InProcessEmitToolchain.Default).WithId("InProcess"),
             ]
             );
             string snapshotDirectory = Path.Combine(Directory.GetCurrentDirectory(), "BenchmarkDotNet.Artifacts", "snapshots");
@@ -43,7 +42,7 @@ namespace BenchmarkDotNet.IntegrationTests
             Output.WriteLine("---------------------------------------------");
             Output.WriteLine("SnapshotDirectory:" + snapshotDirectory);
             var snapshots = Directory.EnumerateFiles(snapshotDirectory)
-                .Where(filePath => Path.GetExtension(filePath).Equals(".dtp", StringComparison.OrdinalIgnoreCase))
+                .Where(filePath => Path.GetExtension(filePath).Equals(".dmw", StringComparison.OrdinalIgnoreCase))
                 .Select(Path.GetFileName!)
                 .OrderBy(fileName => fileName)
                 .ToList();
@@ -51,18 +50,27 @@ namespace BenchmarkDotNet.IntegrationTests
             foreach (var snapshot in snapshots)
                 Output.WriteLine("* " + snapshot);
 
-            Assert.Single(snapshots);
-            // Assert.Equal(2, snapshots.Count);
+            Assert.Equal(2, snapshots.Count);
+            // Assert.Equal(4, snapshots.Count);
         }
 
-        [DotTraceDiagnoser]
+        [DotMemoryDiagnoser]
         public class Benchmarks
         {
             [Benchmark]
-            public int Foo()
+            public int Foo0()
             {
                 var list = new List<object>();
-                for (int i = 0; i < 1000000; i++)
+                for (int i = 0; i < 1000; i++)
+                    list.Add(new object());
+                return list.Count;
+            }
+
+            [Benchmark]
+            public int Foo1()
+            {
+                var list = new List<object>();
+                for (int i = 0; i < 1000; i++)
                     list.Add(new object());
                 return list.Count;
             }
